requirements.txt
akshare>=1.13.95
pandas>=2.0.0
PyYAML>=6.0
feedgen>=0.9.0
requests>=2.31.0
